Move useLocation to jsx-helpers
