+++ src/app/components/shell/header/main-menu/login-menu/login-menu.js
@@ -1,25 +1,10 @@
 import settings from 'settings';
 import React, {useState, useEffect} from 'react';
-import {useToggle} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import {useToggle, useLocation} from '~/components/jsx-helpers/jsx-helpers.jsx';
 import {useUserModel} from '~/models/usermodel';
 import userModelBus from '~/models/usermodel-bus';
 import linkHelper from '~/helpers/link';
 import Dropdown, {MenuItem} from '../dropdown/dropdown';
-
-function useLocation() {
-    const [location, setLocation] = useState(window.location.href);
-
-    useEffect(() => {
-        const updateLocation = () => {
-            setLocation(window.location.href);
-        };
-
-        window.addEventListener('navigate', updateLocation);
-        return () => window.removeEventListener('navigate', updateLocation);
-    }, []);
-
-    return location;
-}
 
 // Slightly hacky; avoiding adding a new item to settings,
 // but probably will do so eventually
